requirements.txt
functions-framework==3.*
requests==2.26.0
python-dotenv==0.19.1
google-generativeai
